creation of backend folder for functions, made file for data calculation functions (analyticsCalculations) and migrated the 3 calculation functions into that file from GymDaily Screen
